frontend/requirements.txt
streamlit
requests
numpy
pillow
matplotlib
